// src/app/shared/models/students/student-subject-paper-code-end-exam-marks.ts
export class StudentSubjectPaperCodeEndExamMarks {
    examinationTypeId?: number;
    examinationId?: number;
    operationalVerticalId?: number
    registrationNumber?: string;
    serialNumber?: number;
    subjectPaperCodeId?: number;
    subjectPaperCodeName?: string;
    subjectId?: number;
    subjectName?: string;
    credits?: 4;
    grade?: string;
    gradePoint?: 6;
    isSubmittedForScrutiny? : boolean;
}