// src/app/shared/models/students/student-subject-paper-code-mid-exam-marks.ts
export class StudentSubjectPaperCodeMidExamMarks {
    registrationNumber?: string;
    serialNumber?: number;
    subjectPaperCodeId?: number;
    subjectPaperCodeName?: string;
    subjectId?: number;
    subjectName?: string;
    totalMarks?: number;
    marksObtained?: number;
    marksWeightage?: number;
}